move all i18n related ui files to one i18n folder
