+++ imports/startup/client/index.js
@@ -1,5 +1,5 @@
 import { SimpleSchema } from 'meteor/aldeed:simple-schema';
 import './startup.js';
-import './messages.js';
+import './i18n/messages.js';
 import './router.js';
 SimpleSchema.debug = true;
